pyproject.toml
[tool.poetry]
name = "calender-backend"
version = "0.1.0"
description = ""
authors = ["'leechanghwan' <'showmev12@gmail.com'>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.4"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.6.0"
django-environ = "^0.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
